Module libpasta::key

The key module is for managing key sources.

The idea is that a running application can dynamically insert keys into the key store, which are used for producing and verifying hashes.

Structs

KEY_STORE
Global key storage

LocalStore
Structure used as a global store for keys.

Traits

Store
A key storage source. Permits retrieving and storing keys.