Trait libpasta::primitives::PrimitiveImpl

pub trait PrimitiveImpl: Debug + Send + Sync {
    fn compute(&self, password: &[u8], salt: &[u8]) -> Vec<u8>;
    fn params_as_vec(&self) -> Vec<(&'static str, String)>;
    fn hash_id(&self) -> Hashes;

    fn verify(&self, password: &[u8], salt: &[u8], hash: &[u8]) -> bool { ... }
}

Trait defining the functionality of a hashing primitive.

Required Methods

fn compute(&self, password: &[u8], salt: &[u8]) -> Vec<u8>

Compute the output of the primitive with input password and salt.

fn params_as_vec(&self) -> Vec<(&'static str, String)>

Output the parameters of the primitive as a list of tuples.

fn hash_id(&self) -> Hashes

Return algorithm type as a MCF-compatible hash identifier.

Provided Methods

fn verify(&self, password: &[u8], salt: &[u8], hash: &[u8]) -> bool

Verify the password and salt against the hash.

In many cases, this just checks whether compute(password, salt) == hash.